Add support for GitHub reusable workflow definition & use (#1867)
* add: GitHub Action workflow trigger workflow_call

* Split GitHub Workflow schema job definitions

There are two flavors of jobs: "normal" jobs, and jobs which call
reusable workflows. Split the two cases using a oneOf schema.

In the previously existing case of a "normal" job, "runs-on" is
required and "uses" is forbidden. In the new case, the inverse is
true: "uses" is required and "runs-on" is forbidden.

The job types are moved into the definitions for the schema, and then
written under the desired `oneOf` block. The "needs" property, shared
between the two, is also moved to a new defintion for reuse.

* Don't require 'description' in workflow_call

For both 'inputs' and 'secrets', the 'description' field is optional,
not required. Correct this error.

* Add test cases for github reusable workflows

* Reusable GitHub Workflows require input 'type'

In reusable GitHub Workflows, which define the `workflow_call`
trigger, the `type` field on inputs is required.

1. Add a negative test case which should fail to validate
2. Mark the field required so that the negative test passes

Co-authored-by: Justin Rittenhouse <jrittenh@nd.edu>