Partial interpolation in GitHub Workflow runs-on (#2312)
GitHub Workflows have an expression syntax covered by a well-defined
regex in the current schema. However, the expression syntax is not
necessarily the full content of a field. In addition to
```
foo: ${{ expr }}
```
the schema should support
```
foo: bar-${{ expr }}
```
in certain contexts.

This PR introduces a new variant of the established regex which allows
for arbitrary prefix and suffix strings. The only requirement is that
a valid expression-syntax-block appear somewhere within the field.

Apply `stringContainingExpressionSyntax` in lieu of `expressionSyntax`
to the `runs-on` field in workflows. This fixes #2311

The `stringContainingExpressionSyntax` may turn out to be desirable in
all locations which use `expressionSyntax`, but this change is not
made here, only in `runs-on`.

Additionally, the `stringContainingExpressionSyntax` block used for
`runs-on` could be made more restrictive. For example, `allOf` could
be used to enforce that if there is a string prefix, it must match one
of the known classes of runners. However, such logic would be
error-prone, and is omitted for now.